Reformatted some stuff
