+++ build_database.py
@@ -2,7 +2,7 @@ from database_interface import *
 from legacy_interface import ask_legacy, LegacyParts, post_scalars
 from datetime import datetime
 
-from sqlalchemy import select, update, and_
+from sqlalchemy import select
 
 import random  # We use random numbers because some the exact numbers of some values shouldn't matter
 import faker
@@ -53,54 +53,6 @@ def build_db():
         session.add(OrderItem(order_id=4, item_id=40, quantity=300))
 
         session.commit()
-
-
-def evaluate_missing_costnweight():
-    with Session(ENGINE) as session:
-        
-        # Update the price and weight of each order item if it's missing.
-        for a in session.execute(select(OrderItem).where(OrderItem.cost.is_(None))).scalars().all():
-            part_weight, part_price = post_scalars(ask_legacy(select(LegacyParts.weight, LegacyParts.price).where(LegacyParts.number == a.item_id)))
-
-            a.cost = part_price * a.quantity
-            a.weight = part_weight * a.quantity
-        
-
-        session.commit()
-    
-    print(*full_table(OrderItem), sep="\n")
-
-
-def update_order_weight():
-    # We do all at once because it's then we know it will always be correct
-    with Session(ENGINE) as session:
-        for a in session.execute(select(Order)).scalars():
-            sum_price = 0
-            sum_weight = 0
-            for b in session.execute(select(OrderItem).where(OrderItem.order_id == a.id)).scalars():
-                part_price, part_weight = post_scalars(ask_legacy(select(LegacyParts.price, LegacyParts.weight).where(LegacyParts.number == b.item_id)))
-                b.cost = part_price * b.quantity
-                b.weight = part_weight * b.quantity
-                sum_price += part_price * b.quantity
-                sum_weight += part_weight * b.quantity
-                # print(b)
-            
-            a.total_cost = sum_price
-            a.total_weight = sum_weight
-
-            fee = session.execute(select(Fees).where(and_(Fees.min_weight <= sum_weight, Fees.max_weight > sum_weight))).scalar()
-            if fee:
-                # print(fee)
-                a.fee_id = fee.id
-                a.total_cost_post_fee = a.total_cost + fee.weight_m * a.total_weight + fee.weight_b
-            else:
-                print("No fee found for order", a.id)
-            print(a)
-        
-        session.commit()
-
-            
-
 
 
 def main():
